requirements.txt
google-cloud-aiplatform
google-generativeai
sympy
pandas
numpy
python-dotenv
jsonschema
matplotlib
seaborn
jupyter
fastapi
uvicorn
python-multipart
